+++ src/bot/commands/slash/index.ts
@@ -2,11 +2,13 @@ import { CogSlash } from "cocoa-discord-utils/slash";
 
 import { aboutme } from "./aboutme";
 import { getstatement } from "./getstatement";
+import { status } from "./status";
 
 export const Cocoa: CogSlash = {
     name: "Cocoa",
     commands: {
         aboutme,
         getstatement,
+        status,
     },
 };